+++ app/src/main/java/com/mbusino/mqttexplorer/MqttExplorerApp.kt
@@ -1,9 +1,20 @@
 package com.mbusino.mqttexplorer
 
 import android.app.Application
+import android.content.Context
 
 class MqttExplorerApp : Application() {
+
+    init {
+        instance = this
+    }
+
     override fun onCreate() {
         super.onCreate()
+    }
+
+    companion object {
+        private lateinit var instance: MqttExplorerApp
+        fun getAppContext(): Context = instance.applicationContext
     }
 }